+++ report-service/src/main/java/com/example/reportservice/controller/VisitAuthRequestController.java
@@ -4,7 +4,7 @@ import com.example.reportservice.common.aop.AdminUser;
 import com.example.reportservice.common.constant.VisitAuthConstant;
 import com.example.reportservice.dto.VisitAuthRequestDetailDTO;
 import com.example.reportservice.dto.VisitAuthRequestResponseDTO;
-import com.example.reportservice.service.VisitAuthRequestService;
+import com.example.reportservice.service.visit_auth.VisitAuthRequestService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Slice;
 import org.springframework.http.HttpStatus;
@@ -44,5 +44,12 @@ public class VisitAuthRequestController {
     @AdminUser
     public ResponseEntity<VisitAuthRequestDetailDTO> getVisitAuthRequest(final @PathVariable int visitAuthId) {
         return ResponseEntity.ok().body(visitAuthRequestService.getVisitAuthRequest(visitAuthId));
+    }
+
+    @PostMapping("/{visitAuthId}/authenticate")
+    @AdminUser
+    public ResponseEntity<Void> authenticateVisitAuthRequest(final @PathVariable int visitAuthId) {
+        visitAuthRequestService.authenticateVisitAuthRequest(visitAuthId);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 }
